Add DOD item about trying to avoid adding 'debt' tasks.
We have had extensive discussions about this topic and in order for us to try to be in agreement about how to go about leaving work behind for future devs it would be good to have this formalised here.

I think the spirit should be that we should *never* add future work when we are building features. We do add future work in maintenance and things we don't anticipate anyways, this is just a way to not add work debt consciously.