Styleguide: casing in shell scripts
Discussion: https://auctionet.slack.com/archives/CF9SAN79V/p1637165952079500

I don't think the env bit is controversial or really needs saying, but let's be extra explicit since we say to make others lowercase.

Did not mention shell variables, which are also typically capitalized, since they can be thought of as a kind of environment variable and we probably don't need to explicitly cover them.

(Going by the definition of env and shell variables here: https://www.digitalocean.com/community/tutorials/how-to-read-and-set-environmental-and-shell-variables-on-linux)